requirements.txt
PyQt6>=6.9.0
Pillow==10.2.0
pyinstaller>=6.0.0 